requirements.txt
streamlit==1.29.0
python-docx==1.1.0
pdfplumber==0.10.3
anthropic==0.7.8
python-dotenv==1.0.0
Pillow==10.1.0
